ResMerger CustomTool Installer
