+++ ResMergerCustomTool/Properties/AssemblyInfo.cs
@@ -1,5 +1,7 @@
 ﻿using System.Reflection;
 using System.Runtime.InteropServices;
+using Microsoft.VisualStudio;
+using ResMergerCustomTool;
 
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
