+++ ResMergerCustomToolInstall/Properties/AssemblyInfo.cs
@@ -1,6 +1,8 @@
 ﻿using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using Microsoft.VisualStudio;
+using ResMergerCustomTool;
 
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
@@ -31,3 +33,6 @@ using System.Runtime.InteropServices;
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("1.0.0.0")]
 [assembly: AssemblyFileVersion("1.0.0.0")]
+
+[assembly: ProvideAssemblyObject(typeof(ResMergerGenerator))]
+[assembly: ProvideGenerator(typeof(ResMergerGenerator), VSConstants.UICONTEXT.CSharpProject_string, Description = "Merges the ResourceDictionaries", GeneratesDesignTimeSource = true)]
