+++ ResMergerCustomToolInstall/ProvideGeneratorAttribute.cs
